+++ common/src/main/java/io/github/flemmli97/simplequests/gui/QuestGui.java
@@ -70,7 +70,7 @@ public class QuestGui extends ServerOnlyScreenHandler<Object> {
     private ItemStack ofQuest(int i, Quest quest, ServerPlayer player) {
         PlayerData data = PlayerData.get(player);
         PlayerData.AcceptType type = data.canAcceptQuest(quest);
-        ItemStack stack = new ItemStack(type == PlayerData.AcceptType.ACCEPT ? Items.PAPER : Items.BOOK);
+        ItemStack stack = type == PlayerData.AcceptType.ACCEPT ? quest.icon.copy() : new ItemStack(Items.BOOK);
         stack.setHoverName(new TextComponent(quest.questTaskString).setStyle(Style.EMPTY.withItalic(false).applyFormat(ChatFormatting.GOLD)));
         ListTag lore = new ListTag();
         if (data.isActive(quest)) {