+++ fabric/src/main/java/io/github/flemmli97/simplequests/fabric/LoaderImpl.java
@@ -16,6 +16,7 @@ import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 
 import java.nio.file.Path;
@@ -32,6 +33,11 @@ public class LoaderImpl implements LoaderHandler {
     @Override
     public ResourceLocation fromEntity(Entity entity) {
         return Registry.ENTITY_TYPE.getKey(entity.getType());
+    }
+
+    @Override
+    public Item fromID(ResourceLocation id) {
+        return Registry.ITEM.get(id);
     }
 
     @Override